@@ -5,7 +5,7 @@ homepage: https://github.com/Abhi011999/flutter_colored_progress_indicators
 repository: https://github.com/Abhi011999/flutter_colored_progress_indicators
 
 environment:
-  sdk: '>=2.12.0-0 <3.0.0'
+  sdk: ">=2.14.1 <3.0.0"
 
 dependencies:
   flutter:
